ENGR00255799 fix usb camera incorrect parameter set.
USB Camera: Gallery force stop and can't open Camera again
if switch front&back Camera.
The root cause is array not zero and accessing array is not
limited to the specified size.

add focal length and sensitive element size for usb camera.
The value is a reference and should be changed according to
the specified usb camera.

Signed-off-by: Xiaowen Liu <b37945@freescale.com>