tinyalsa: allow HDMI to be used as primary output device
This patch allows the HDMI audio driver (imx-hdmi-soc) to be
used as the primary audio output if the Android system property
"sys.hdmi" is set to 1.

Note that this is not completely dynamic, because the implementation
checks for the property when opening the audio device. In other words,
you can't switch in the middle of audio playback.

You can switch the value of the property from 0 to 1 at run-time,
but the expected use-case is that the audio routing will be set
at boot time based on what's connected to the board.

Note that you can add a line into /system/etc/build.prop, but
remember to set the mode back to 644 (chmod 644 /system/build.prop)
or your system will become un-bootable.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>